Fix issue with gradient computation.
We missed parts of the gradient coefficients c'_k for
all phi_k with k at the upper boundary of the basis
shape K. These are high-k contributions which often
have negligible coefficients. But in the example of a
Gaussian phi_0 and |K| = 1 we get zero gradient and
zero kinetic energy in case p = 0.

Note: Algorithm 9 on p. 64 of MT *is correct* only the
implementation was wrong.