Optionally disable computation of Q0 and P0
For some potentials the Hessian matrix vanishes in the minimum.
This will give issues when computing Q0 from H^(-1/4). This patch
disables computation of these parameters by the switch '--noPQ'.
If present, the values will be read from the configuration file
variables 'Q0' and 'P0'. Otherwise they are set to unity.